requirements-dev.txt
# Runtime dependencies
fastapi
uvicorn[standard]
httpx
pyyaml
prometheus-client
starlette
pydantic

# Development dependencies (to be removed for production)
pytest
pytest-asyncio
pytest-httpx
respx

